Backend/requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.1
Pillow==9.3.0
